requirements.txt
nltk
requests
beautifulsoup4
google
numpy
pandas
scikit-learn
newsapi-python
praw
tweepy
readability-lxml
textblob
black
lxml_html_clean